+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## 0.33.0
+* Add export ideas-with-arguments
+
 ## 0.32.0
 * Add automaticallyUpdateStatus to site config
 * Add 'Project has ended' to site configuration